requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.1


djangorestframework
#pillow==8.4
django-cors-headers
drf-yasg
coreapi
setuptools
# pkg-resources==0.0.0